# service/internal/httpserver/server.go
// Package httpserver wires the HTTP API, static assets and server lifecycle.
package httpserver

import (
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"io/fs"
	"log/slog"
	"maps"
	"net/http"
	"sync"
	"sync/atomic"
	"time"

	"github.com/danielgtaylor/huma/v2"

	"github.com/s-frei/rezepte/service/internal/config"
)

// Server owns the mux, the huma API and the http.Server lifecycle.
type Server struct {
	cfg       config.Config
	logger    *slog.Logger
	mux       *http.ServeMux
	api       huma.API
	specGuard func(http.Handler) http.Handler
	preview   PreviewFunc
	headers   ShellHeaderFunc
	version   string
}

// Option configures a Server at construction time. Options run after the
// huma API is created but before New returns, so an Option such as
// WithAPIMiddleware is guaranteed to apply before the caller can register
// any operation on the returned Server.
type Option func(*Server)

// WithAPIMiddleware installs an API-wide huma middleware. mw receives the
// server's huma.API (useful for writing errors with huma.WriteErr) and
// returns the actual middleware function.
//
// Because New applies every Option before returning, a middleware installed
// this way is structurally guaranteed to run for every operation registered
// afterwards - there is no window in which an operation could be registered
// ahead of it, unlike calling api.UseMiddleware from arbitrary caller code.
func WithAPIMiddleware(mw func(api huma.API) func(huma.Context, func(huma.Context))) Option {
	return func(s *Server) {
		s.api.UseMiddleware(mw(s.api))
	}
}

// WithSpecGuard wraps the routes huma registers for the contract itself -
// the OpenAPI document, the JSON schemas and the Scalar docs page.
//
// Those are not huma operations: huma hangs them straight off the mux, so
// the middleware WithAPIMiddleware installs never runs for them and their
// Security is empty. Guarding them therefore has to happen here, in front
// of the mux, rather than in the operation chain. mw is an ordinary
// net/http middleware - auth.RequireAuthOrLogin, which redirects a
// browser to the login form and answers everything else with a 401; see
// docs/memory/content/features/users-and-auth.mdx.
func WithSpecGuard(mw func(http.Handler) http.Handler) Option {
	return func(s *Server) {
		s.specGuard = mw
	}
}

// WithLinkPreview installs the function that gives a page its own link
// preview in the app shell - a shared recipe's title and cover. Without it
// every page carries the app's own. See docs/memory/content/features/link-previews.mdx.
func WithLinkPreview(f PreviewFunc) Option {
	return func(s *Server) { s.preview = f }
}

// WithShellHeaders installs the function that adjusts the app shell's
// response headers for the page a request asks for - the public share page
// sends noindex and no-referrer, for one. Without it every page gets the
// shell's own headers.
func WithShellHeaders(f ShellHeaderFunc) Option {
	return func(s *Server) { s.headers = f }
}

// WithVersion sets the build version /healthz reports and the OpenAPI
// document carries as info.version. Without it both say "dev", which is what
// a plain `go build` produces.
func WithVersion(v string) Option {
	return func(s *Server) { s.version = v }
}

// WithSecuritySchemes declares the OpenAPI security schemes the operations
// refer to. huma.DefaultConfig declares none, so without this the emitted
// document names schemes it never defines and the docs page offers no way to
// authorize. The map comes from the auth package, which owns the names -
// httpserver stays ignorant of how authentication works, exactly as it does
// for WithAPIMiddleware.
func WithSecuritySchemes(schemes map[string]*huma.SecurityScheme) Option {
	return func(s *Server) {
		components := s.api.OpenAPI().Components
		if components.SecuritySchemes == nil {
			components.SecuritySchemes = map[string]*huma.SecurityScheme{}
		}
		maps.Copy(components.SecuritySchemes, schemes)
	}
}

// errorLogger holds the logger the huma error hook below logs to. New stores
// its logger here; when multiple Servers exist, the last call to New wins
// for every server's errors, since the hook is a single package-level
// override of huma.NewErrorWithContext. This is fine in this process, which
// only ever runs one Server, and is acceptable in tests, which each build
// their own short-lived Server sequentially.
var errorLogger atomic.Pointer[slog.Logger]

// installErrorHookOnce guards the one-time capture of huma's original
// NewErrorWithContext and installation of the replacement below.
var installErrorHookOnce sync.Once

// installErrorHook overrides huma.NewErrorWithContext - the constructor huma
// calls both for handler errors and from huma.WriteErr - so that a plain
// error returned from a handler (mapped by huma to a 500) never leaks its
// message (which may contain internal details such as SQL driver errors) to
// the client. Status codes below 500 keep huma's normal behavior via the
// captured original constructor.
func installErrorHook() {
	installErrorHookOnce.Do(func() {
		original := huma.NewErrorWithContext
		huma.NewErrorWithContext = func(ctx huma.Context, status int, msg string, errs ...error) huma.StatusError {
			if status < http.StatusInternalServerError {
				return original(ctx, status, msg, errs...)
			}
			if logger := errorLogger.Load(); logger != nil {
				logger.Error("internal error",
					"operation", ctx.Operation().OperationID,
					"method", ctx.Method(),
					"path", ctx.URL().Path,
					"err", errors.Join(errs...))
			}
			return &huma.ErrorModel{
				Title:  http.StatusText(status),
				Status: status,
				Detail: "internal error",
			}
		}
	})
}

// New builds a server serving the API under /api/v1 and static assets from
// fsys. Options run after the huma API is created and before any operation
// can be registered on it; use WithAPIMiddleware to install API-wide
// middleware (such as auth.Middleware) with that guarantee.
func New(cfg config.Config, logger *slog.Logger, static fs.FS, opts ...Option) *Server {
	mux := http.NewServeMux()
	s := &Server{cfg: cfg, logger: logger, mux: mux, api: newAPI(mux), version: defaultVersion}

	installErrorHook()
	errorLogger.Store(logger)

	for _, opt := range opts {
		opt(s)
	}
	// The API is versioned with the application: one number, bumped by the
	// release tag, never by hand. See docs/memory/content/architecture/api.mdx.
	s.api.OpenAPI().Info.Version = s.version

	// The version is public here by decision: it is the one place an operator
	// can read what a running instance is without shell access to it.
	mux.HandleFunc("GET /healthz", func(w http.ResponseWriter, _ *http.Request) {
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		_ = json.NewEncoder(w).Encode(map[string]string{"status": "ok", "version": s.version})
	})
	// huma's own operations register more specific patterns (e.g.
	// "GET /api/v1/openapi.json") which take precedence over these catch-alls.
	mux.Handle("/api", notFound("no such API route"))
	mux.Handle("/api/", notFound("no such API route"))
	// Nothing is published under /.well-known/. An MCP client that gets a
	// 401 from /mcp looks there for OAuth metadata, and the SPA's index.html
	// with a 200 would read as a broken document rather than "none here".
	mux.Handle("/.well-known/", notFound("no such document"))
	mux.Handle("/", SPAHandler(static, cfg.SecureCookies, s.preview, s.headers))
	return s
}

// notFound answers a route that must not fall through to the SPA with an
// RFC 9457 problem+json 404 instead of the stdlib's plain-text one.
func notFound(detail string) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
		WriteProblem(w, http.StatusNotFound, detail)
	})
}

// WriteProblem answers with an RFC 9457 problem+json body in the shape huma
// gives its own errors, for the responses written outside a huma operation.
func WriteProblem(w http.ResponseWriter, status int, detail string) {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(huma.ErrorModel{
		Title:  http.StatusText(status),
		Status: status,
		Detail: detail,
	})
}

// publicError is huma's ErrorModel under a type of its own. huma's schema
// link transformer recognizes a response body by its Go type, and a bare
// ErrorModel would get a "$schema" key and a Link header pointing at a
// schema route that needs a session.
type publicError struct{ huma.ErrorModel }

// PublicNotFound is the 404 of an operation anyone may call without a
// session: the problem+json body huma gives, minus the schema link a
// stranger could not follow.
func PublicNotFound(detail string) error {
	return &publicError{huma.ErrorModel{
		Title:  http.StatusText(http.StatusNotFound),
		Status: http.StatusNotFound,
		Detail: detail,
	}}
}

// API exposes the huma API so feature packages can register operations.
func (s *Server) API() huma.API { return s.api }

// Handle registers h on the server's mux for a route that is not a huma
// operation (Phase 4's image files). Go 1.22 mux precedence picks the most
// specific matching pattern, so a pattern such as
// "GET /images/{recipeId}/{imageId}/{file}" wins over the "/" SPA
// catch-all New installed, whatever the registration order. The handler is
// still wrapped by Handler()'s Origin check and request logging.
func (s *Server) Handle(pattern string, h http.Handler) {
	s.mux.Handle(pattern, h)
}

// Handler returns the root handler with Origin check and request logging.
func (s *Server) Handler() http.Handler { return s.logRequests(checkOrigin(s.guardSpec(s.mux))) }

// guardSpec routes the contract paths through the middleware WithSpecGuard
// installed and lets everything else reach the mux untouched. With no guard
// configured the mux is returned as it is, so a Server built without the
// option behaves exactly as before.
func (s *Server) guardSpec(next http.Handler) http.Handler {
	if s.specGuard == nil {
		return next
	}
	guarded := s.specGuard(next)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if isSpecRoute(r.URL.Path) {
			guarded.ServeHTTP(w, r)
			return
		}
		next.ServeHTTP(w, r)
	})
}

// Run serves until ctx is canceled, then shuts down gracefully.
func (s *Server) Run(ctx context.Context) error {
	srv := &http.Server{
		Addr:              s.cfg.Addr,
		Handler:           s.Handler(),
		ReadHeaderTimeout: 10 * time.Second,
	}
	errc := make(chan error, 1)
	go func() {
		s.logger.Info("listening", "addr", s.cfg.Addr)
		errc <- srv.ListenAndServe()
	}()
	select {
	case err := <-errc:
		if errors.Is(err, http.ErrServerClosed) {
			return nil
		}
		return fmt.Errorf("listen on %s: %w", s.cfg.Addr, err)
	case <-ctx.Done():
		shutdownCtx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
		defer cancel()
		if err := srv.Shutdown(shutdownCtx); err != nil {
			return fmt.Errorf("shutdown: %w", err)
		}
		return nil
	}
}

func (s *Server) logRequests(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		start := time.Now()
		rec := &statusRecorder{ResponseWriter: w, status: http.StatusOK}
		next.ServeHTTP(rec, r)

		level := slog.LevelDebug
		switch {
		case rec.status >= http.StatusInternalServerError:
			level = slog.LevelError
		case rec.status >= http.StatusBadRequest:
			level = slog.LevelWarn
		}
		s.logger.Log(r.Context(), level, "request",
			"method", r.Method, "path", r.URL.Path,
			"status", rec.status, "duration", time.Since(start))
	})
}

type statusRecorder struct {
	http.ResponseWriter
	status int
}

func (r *statusRecorder) WriteHeader(code int) {
	r.status = code
	r.ResponseWriter.WriteHeader(code)
}

// Unwrap exposes the underlying writer so http.ResponseController can reach
// Flush, Hijack and deadline support through the logging middleware.
func (r *statusRecorder) Unwrap() http.ResponseWriter { return r.ResponseWriter }